Rewrote test cases and fixed loading.
I made the loading code use the unionRangeMerge code that we just wrote. Now loading is as
assured as our union code is and it is super fast and simple. I really like the look of
our code now and am more than happy with the results. Our test cases are also helping to
spot our errors as we go and that is fantastic.